-merged three files into one -grouped tasks
